Initial commit!
This server can accept multiple connections, receive pings and respond
with pongs, respond with multiplayer list status, and accept logins. The
logins will not get very far, however, and are simply responded to with
a (Login) disconnect packet.

The `Serializer` directory creates a binary to aid in the creation of
packet (de)serializers. A simple JSON format is used, and the only
manual requirement is adding the packet ID definition into the proper
enum inside of `Packet.h`.

The `LibMinecraft/Chat` directory is used to create/parse chat
components. These are currently used in the Disconnect packet. Most
crucial chat component features can be represented currently -- however
the API itself is not the greatest and is quite verbose.