LibMinecraft: Introduce parsing of Sponge schematic format
Sponge schematics are used by WorldEdit, and probably other software,
and the specification is quite nice:
https://github.com/SpongePowered/Schematic-Specification

To do this, we had to introduce a `ResourceLocation` type (also known as
a namespaced key). This has a namespace, and a path to a resource.
It can consist of lowercase characters, numbers, underscores, hyphens
and dots. In the path, a forward slash can represent a "directory
separator". Minecraft prefers the use of `snake_case` in these strings.
Minecraft uses the standard namespace of `minecraft`. A block might be
`minecraft:grass_block`, an item might be `minecraft:stick`, etc. This
goes for _any_ resource. Resources not originating from Minecraft
_should_ use a different namespace. The Dark Matter item from ProjectE
mod is known as `projecte:dark_matter`

Blocks don't only have a resource location, but also have a key-value
style list of properties attached to them. We implement this as
`BlockState`. Although we don't know what keys and values mean, or
the default or possible keys on different blocks -- we can at least
store them properly.